+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.1.2 (01-05-2024)
+
+- Added workflow transformer for STAC collections
+  - Generate STAC collection definition for workflows
+  - Add required fields for a collection
+  - Scrape CWL script (if provided) to fill gaps in definition
+  - Generate workflow ID (uuid) if not provided
+  - Upload transformed workflow to S3 bucket
+- Updated logging to streamline debug logs
+
 ## v0.1.1 (25-04-2024)
 
 - Added workflows